Split routes.py into a rendering/redirection file and a file for everything else
